# static/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: "CSI Korika"
  description: "API documentation for CSI Korika Malaria Monitoring System"
basePath: "/"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\""
paths:
  /signup:
    post:
      tags:
        - "Authentication"
      summary: "User Registration"
      description: "Register a new user account"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User registration information"
          required: true
          schema:
            type: "object"
            required:
              - "email"
              - "password"
            properties:
              email:
                type: "string"
                format: "email"
                example: "user@example.com"
              password:
                type: "string"
                format: "password"
                example: "securepassword123"
              full_name:
                type: "string"
                example: "John Doe"
              phone_number:
                type: "string"
                example: "+6281234567890"
              address_1:
                type: "string"
                example: "Jl. Sukasari No. 123"
              address_2:
                type: "string"
                example: "Kelurahan Sukajadi"
              access_level:
                type: "string"
                example: "user"
                enum: ["user", "admin"]
      responses:
        201:
          description: "User successfully registered"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              access_token:
                type: "string"
              user:
                type: "object"
              success:
                type: "boolean"
        400:
          description: "Invalid input or user already exists"
          schema:
            type: "object"
            properties:
              error:
                type: "string"
  /signin:
    post:
      tags:
        - "Authentication"
      summary: "User Login"
      description: "Authenticate user and receive access token"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User login credentials"
          required: true
          schema:
            type: "object"
            required:
              - "email"
              - "password"
            properties:
              email:
                type: "string"
                format: "email"
                example: "user@example.com"
              password:
                type: "string"
                format: "password"
                example: "securepassword123"
      responses:
        200:
          description: "Login successful"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              access_token:
                type: "string"
              user:
                type: "object"
              success:
                type: "boolean"
        401:
          description: "Invalid credentials"
          schema:
            type: "object"
            properties:
              error:
                type: "string"
  /get-provinces:
    get:
      tags:
        - "Data"
      summary: "Get List of Provinces"
      description: "Retrieve list of all provinces available in the system"
      produces:
        - "application/json"
      responses:
        200:
          description: "List of provinces successfully retrieved"
          schema:
            type: "object"
            properties:
              data:
                type: "array"
                items:
                  type: "string"
              success:
                type: "boolean"
        401:
          description: "Unauthorized"
        500:
          description: "Server error"
      security:
        - Bearer: []
  /get-raw-data:
    get:
      tags:
        - "Data"
      summary: "Get Raw Malaria Data"
      description: "Retrieve raw malaria data with optional province filter"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "province"
          type: "string"
          description: "Province name to filter data (use 'TOTAL' for all provinces)"
          required: false
      responses:
        200:
          description: "Raw data successfully retrieved"
          schema:
            type: "object"
            properties:
              data:
                type: "array"
                items:
                  type: "object"
              success:
                type: "boolean"
        401:
          description: "Unauthorized"
        500:
          description: "Server error"
      security:
        - Bearer: []
  /get-aggregate-data:
    get:
      tags:
        - "Data"
      summary: "Get Aggregated Malaria Data"
      description: "Retrieve aggregated malaria data with filters"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "province"
          type: "string"
          description: "Province name to filter data (use 'TOTAL' for nationwide data)"
          required: true
        - in: "query"
          name: "year"
          type: "integer"
          description: "Year to filter data"
          required: false
        - in: "query"
          name: "month"
          type: "integer"
          description: "Month to filter data (1-12)"
          required: false
      responses:
        200:
          description: "Aggregated data successfully retrieved"
          schema:
            type: "object"
            properties:
              data:
                type: "array"
                items:
                  type: "object"
              success:
                type: "boolean"
        400:
          description: "Missing required parameters"
        401:
          description: "Unauthorized"
        500:
          description: "Server error"
      security:
        - Bearer: []
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      email:
        type: "string"
      full_name:
        type: "string"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
      phone_number:
        type: "string"
      address_1:
        type: "string"
      address_2:
        type: "string"
      access_level:
        type: "string"
        enum: ["user", "admin"]